+++ packages/zefyr/lib/src/widgets/editable_text_block.dart
@@ -7,6 +7,7 @@ import 'controller.dart';
 import 'cursor.dart';
 import 'editable_text_line.dart';
 import 'editor.dart';
+import 'link.dart';
 import 'text_line.dart';
 import 'theme.dart';
 
@@ -20,8 +21,10 @@ class EditableTextBlock extends StatelessWidget {
   final Color selectionColor;
   final bool enableInteractiveSelection;
   final bool hasFocus;
-  final EdgeInsets? contentPadding;
   final ZefyrEmbedBuilder embedBuilder;
+  final LinkActionPicker linkActionPicker;
+  final ValueChanged<String?>? onLaunchUrl;
+  final EdgeInsets? contentPadding;
 
   const EditableTextBlock({
     Key? key,
@@ -35,6 +38,8 @@ class EditableTextBlock extends StatelessWidget {
     required this.enableInteractiveSelection,
     required this.hasFocus,
     required this.embedBuilder,
+    required this.linkActionPicker,
+    this.onLaunchUrl,
     this.contentPadding,
   }) : super(key: key);
 
@@ -70,7 +75,11 @@ class EditableTextBlock extends StatelessWidget {
           devicePixelRatio: MediaQuery.of(context).devicePixelRatio,
           body: TextLine(
             node: line,
+            readOnly: readOnly,
+            controller: controller,
             embedBuilder: embedBuilder,
+            linkActionPicker: linkActionPicker,
+            onLaunchUrl: onLaunchUrl,
           ),
           cursorController: cursorController,
           selection: selection,